+++ unix/Xvnc/programs/Xserver/damageext/damageext.c
@@ -1,6 +1,8 @@
 /*
  * Copyright © 2002 Keith Packard
  * Copyright 2013 Red Hat, Inc.
+ * Copyright © 2023 Kasm
+ * Copyright © 2024 D. R. Commander
  *
  * Permission to use, copy, modify, distribute, and sell this software and its
  * documentation for any purpose is hereby granted without fee, provided that
@@ -29,6 +31,10 @@
 #include "damagestr.h"
 #include "protocol-versions.h"
 #include "extinit.h"
+
+#if defined(DRI3) && defined(TURBOVNC)
+extern void xvnc_sync_dri3_textures(void);
+#endif
 
 #ifdef PANORAMIX
 #include "panoramiX.h"
@@ -96,6 +102,10 @@ DamageExtNotify(DamageExtPtr pDamageExt, BoxPtr pBoxes, int nBoxes)
     int i, x, y, w, h;
 
     damageGetGeometry(pDrawable, &x, &y, &w, &h);
+
+#if defined(DRI3) && defined(TURBOVNC)
+    xvnc_dri3_sync_pixmaps_to_bos();
+#endif
 
     UpdateCurrentTimeIf();
     ev = (xDamageNotifyEvent) {